move models into their own folder
fix use statements

live on dev server
