--- laravel/app/Google.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Google extends Model
-{
-    public static function updateSheets($date)
-    {
-        
-    }
-}
