+++ laravel/app/Http/Controllers/admin/DBEditorController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\admin;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use DB;
+use Illuminate\Support\Facades\DB;
 
 class DBEditorController extends Controller
 {
