+++ laravel/app/Http/Middleware/CheckResident.php
@@ -4,8 +4,8 @@ namespace App\Http\Middleware;
 
 use Closure;
 
-use App\Resident;
-use App\Admin;
+use App\Models\Resident;
+use App\Models\Admin;
 
 class CheckResident
 {
