+++ laravel/app/Models/ScheduleData.php
@@ -1,17 +1,18 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class EvaluateData extends Model
+class ScheduleData extends Model
 {
+    
     /**
      * The table associated with the model.
      *
      * @var string
      */
-    protected $table = 'evaluation_data';
+    protected $table = 'schedule_data';
 
     /**
      * Indicates if the model should be timestamped.
