+++ laravel/app/Models/ScheduleDataStatic.php
@@ -1,17 +1,18 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Announcements extends Model
+class ScheduleDataStatic extends Model
 {
+    
     /**
      * The table associated with the model.
      *
      * @var string
      */
-    protected $table = 'announcements';
+    protected $table = 'schedule_data_static';
 
     /**
      * Indicates if the model should be timestamped.
@@ -19,5 +20,8 @@ class Announcements extends Model
      * @var bool
      */
     public $timestamps = true;
-
+	
+	public function getTableColumns() {
+        return $this->getConnection()->getSchemaBuilder()->getColumnListing($this->getTable());
+    }
 }
